requirements.txt
chess==1.10.0
langchain==0.2.4
langchain_core==0.2.6
langchain_google_genai==1.0.6
langchain_groq==0.1.5
langchain_openai==0.1.8
python-dotenv==1.0.1
pygame